% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_results.R
\name{compare_new_diet}
\alias{compare_new_diet}
\title{Compare new diet with current diet}
\usage{
compare_new_diet(data_new_diet, data_current_diet)
}
\arguments{
\item{data_new_diet}{New diet intake}

\item{data_current_diet}{Current diet that has food name and average intake}
}
\value{
A dataframe with food names, current and new diet intake and comparsion metrics
}
\description{
Compare new and current diet in terms of absolute and relative (percent) change
}
