% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_constraints.R
\name{compute_constr}
\alias{compute_constr}
\title{Compute constraint values for outcomes}
\usage{
compute_constr(data_total_contrib, data_constr_coef)
}
\arguments{
\item{data_total_contrib}{A dataframe where total contribution (raw or standardised) for each \code{tag_outcome} is stored}

\item{data_constr_coef}{A dataframe of lower and upper constraint coefficients for each \code{tag_outcome}}
}
\value{
A dataframe of calculated constraint values
}
\description{
Compute constraint values for each tag_outcomes, used for the optimisation algorithm.
The computation is based on constraint coefficients, where the basis to multiply the coefficients are total contribution (either raw or standardized).
}
\examples{
diet_selected <- select_diet(data_diet = all_diet,
                      tag_food = c('Bread', 'Vegetables', 'Red meat'))

cpu_selected <- select_perunit(data_perunit_contrib = contrib_per_unit, 
                        tag_food = c('Bread', 'Vegetables', 'Red meat'), 
                        tag_outcome = c('energy', 'protein', 'ghge'))

tc <- compute_total_contrib(data_diet = diet_selected, 
                      data_perunit_contrib = cpu_selected)
constr_coef_df <- set_constr_coef(tag_outcome = c('energy', 'protein', 'ghge'), 
                                  coef_lwr = rep(0.9, 3), 
                                  coef_upr = rep(1.0, 3))

# reduce ghge to 0.9
constr_coef_df_red <- reduce_constr(data_constr_coef = constr_coef_df, 
                                    tag_outcome_reduce = 'ghge', 
                                    coef_reduce = 0.9)
compute_constr(data_total_contrib = tc$total_contrib, 
               data_constr_coef = constr_coef_df_red)
}
