% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_constraints.R
\name{compute_stdcoef}
\alias{compute_stdcoef}
\title{Compute standardise coefficients for contribution per unit}
\usage{
compute_stdcoef(data_perunit_contrib, method = "sd")
}
\arguments{
\item{data_perunit_contrib}{A dataframe where contribution per unit for each tag_outcome is stored.}

\item{method}{Method to compute the coefficient. So far only \code{method = 'sd'} is implemented.}
}
\value{
A list of outputs:
\itemize{
\item \code{std_coef} stores the standardization coefficients to scale the contribution per unit for each tag_outcome
\item \code{method} returns the method used for the computation
}
}
\description{
Compute standardise coefficients for contribution per unit for each \code{tag_outcome}.
}
\examples{
cpu_selected <- select_perunit(data_perunit_contrib = contrib_per_unit, 
                        tag_food = c('Bread', 'Vegetables', 'Red meat'), 
                        tag_outcome = c('energy', 'protein', 'ghge'))
compute_stdcoef(data_perunit_contrib = cpu_selected)
}
