% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_constraints.R
\name{compute_total_contrib}
\alias{compute_total_contrib}
\title{Compute total contribution for outcomes}
\usage{
compute_total_contrib(data_diet, data_perunit_contrib)
}
\arguments{
\item{data_diet}{A dataframe where diet intake information is stored}

\item{data_perunit_contrib}{A dataframe where contribution per unit for each tag_outcome is stored. This could also be the standardized contribution per unit.}
}
\value{
A list of outputs:
\itemize{
\item \code{total_contrib} stores the total contribution of the current diet
\item \code{tag_food} stores the names of foods
\item \code{tag_outcome} stores the names of outcomes
}
}
\description{
Compute total contribution of the current diet regarding each outcome of interest.
}
\examples{
diet_selected <- select_diet(data_diet = all_diet,
                      tag_food = c('Bread', 'Vegetables', 'Red meat'))

cpu_selected <- select_perunit(data_perunit_contrib = contrib_per_unit, 
                        tag_food = c('Bread', 'Vegetables', 'Red meat'), 
                        tag_outcome = c('energy', 'protein', 'ghge'))

compute_total_contrib(data_diet = diet_selected, 
                      data_perunit_contrib = cpu_selected)
}
