% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_contrib}
\alias{data_contrib}
\title{Demo data: new diet contribution to nutrients and env impact}
\format{
\describe{
\item{tag_outcome}{Nutrient and env impact name}
\item{total_contrib_raw}{}
\item{min}{}
\item{max}{}
\item{tc_new_diet}{}
\item{is_ok}{}
\item{dev_if_not_ok}{}
}
}
\usage{
data_contrib
}
\description{
Demo data for visualization development purposes.
}
\keyword{datasets}
