% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_newdiet}
\alias{data_newdiet}
\title{Demo data: new diet comparison}
\format{
\describe{
\item{food_name}{Food name}
\item{current}{}
\item{new}{}
\item{absolute_change}{}
\item{percent_change}{}
\item{diet_bound_lwr}{}
\item{diet_bound_upr}{}
}
}
\usage{
data_newdiet
}
\description{
Demo data for visualization development purposes.
}
\keyword{datasets}
