% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_algo.R
\name{f_make_constraint_function}
\alias{f_make_constraint_function}
\title{Make inequality constraint functions based on constraint values}
\usage{
f_make_constraint_function(constraint_values, tag_outcomes)
}
\arguments{
\item{constraint_values}{A list of constraint values organized by tag_outcome. Output from values_by_tag_outcome()}

\item{tag_outcomes}{A vector of tag_outcome of interest}
}
\value{
A function to be used in the optimization
}
\description{
Make inequality constraint functions, used as inputs for the optimization.
So far only offers possibility for 7 tag_outcomes; but shall expand.
}
