% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_algo.R
\name{f_make_objective_function}
\alias{f_make_objective_function}
\title{Make objective function based on current diet}
\usage{
f_make_objective_function(diet0, method = "ss")
}
\arguments{
\item{diet0}{A vector of diet intake}

\item{method}{Specify method to use to compute the objective function values. So far only 'ss' (sum of squares) is implemented.}
}
\value{
A function to be used in the optimization
}
\description{
Make objective function based on current diet, used as inputs for the optimization.
}
\examples{
# a demo example with 3 foods intakes
funobj <- f_make_objective_function(diet0 = c(187, 200, 55))
# to evaluate a new diet
funobj(x = c(187, 200, 50))
}
