% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_algo.R
\name{find_new_diet}
\alias{find_new_diet}
\title{New diet discovery}
\usage{
find_new_diet(
  diet0,
  diet0_upr,
  diet0_lwr,
  tag_outcomes,
  constraint_val,
  print_runtime = T
)
}
\arguments{
\item{diet0}{A vector of current diet intake}

\item{diet0_upr}{A vector of values that set the upper bound of diet search}

\item{diet0_lwr}{A vector of values that set the lower bound of diet search}

\item{tag_outcomes}{A vector of tag_outcome of interest}

\item{constraint_val}{A list of constraint values that match tag_outcomes}

\item{print_runtime}{Whether to print runtime. Default is True.}
}
\value{
A list of results
}
\description{
Find new diet combination that satisfies a set of nutrition and environmental impact constraints.
}
