% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{foodname_group}
\alias{foodname_group}
\title{Macro food group (used for plotting purpose)}
\format{
\describe{
\item{group_macro}{7 macro food groups}
\item{food_name}{28 food subgroups}
}
}
\usage{
foodname_group
}
\description{
This dataset contains the information for bigger food groups. We have 7 big groups, and 28 sub-groups.
}
\keyword{datasets}
