% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstfunc.R
\name{fplaceholder}
\alias{fplaceholder}
\title{A placeholder function}
\usage{
fplaceholder(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
another numeric value
}
\description{
A placeholder function
}
\examples{
fplaceholder(x = 2)
}
