% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_results.R
\name{plot.diet_comparison_gram}
\alias{plot.diet_comparison_gram}
\title{Plotting function for diet comparison summary}
\usage{
\method{plot}{diet_comparison_gram}(plot_obj, title_text, axis_x_text, axis_y_text)
}
\arguments{
\item{plot_obj}{an object of class <diet_comparison_percent>}

\item{title_text}{Title for the plot}

\item{axis_x_text}{X axis title}

\item{axis_y_text}{Y axis title}
}
\value{
A plot
}
\description{
Plotting function for diet comparison summary
}
\examples{
library(ggplot2)
ddd <- prep_diet_comparison_gram(data_dietsummary = data_newdiet)
plot(plot_obj = ddd,
                          title_text = 'New diet',
                         axis_x_text = 'Food groups',
                          axis_y_text = 'Intake (grams)')
}
