% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_results.R
\name{plot_diet_comparison_percent}
\alias{plot_diet_comparison_percent}
\title{Plotting function for diet comparison summary}
\usage{
plot_diet_comparison_percent(plot_obj, title_text, axis_x_text, axis_y_text)
}
\arguments{
\item{plot_obj}{an object of class <diet_comparison_percent>}

\item{title_text}{Title for the plot}

\item{axis_x_text}{X axis title}

\item{axis_y_text}{Y axis title}
}
\value{
A plot
}
\description{
Plotting function for diet comparison summary
}
\examples{
ddd <- prep_diet_comparison_percent(data_dietsummary = data_newdiet)
plot_diet_comparison_percent(plot_obj = ddd,
                          title_text = 'Percent change',
                         axis_x_text = 'Food groups',
                          axis_y_text = 'Percent')
}
