% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_results.R
\name{prep_contrib}
\alias{prep_contrib}
\title{Summary table preparation for fast checking constraint fulfilment}
\usage{
prep_contrib(data_contrib, demo = F)
}
\arguments{
\item{data_contrib}{summary table from the results}

\item{demo}{T or F, is this for demonstration}
}
\value{
a dataframe
}
\description{
Summary table preparation for fast checking constraint fulfilment
}
\examples{
dtt <- prep_contrib(data_contrib = data_contrib, demo = T)
}
