% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_results.R
\name{prep_diet_comparison_percent}
\alias{prep_diet_comparison_percent}
\title{Make diet comparison summary table (percent change)}
\usage{
prep_diet_comparison_percent(data_dietsummary)
}
\arguments{
\item{data_dietsummary}{a dataframe that contains x,y,z}
}
\value{
a list of items (pending description)
}
\description{
Make diet comparison summary table (percent change)
}
\examples{
ddd <- prep_diet_comparison_percent(data_dietsummary = data_newdiet)
}
