% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_results.R
\name{prep_validate_table}
\alias{prep_validate_table}
\title{Summary table preparation for fast checking constraint fulfilment}
\usage{
prep_validate_table(data_validate_diet)
}
\arguments{
\item{data_contrib}{summary table from the results}

\item{demo}{T or F, is this for demonstration}
}
\value{
a dataframe
}
\description{
Summary table preparation for fast checking constraint fulfilment
}
