% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_constraints.R
\name{reduce_constr}
\alias{reduce_constr}
\title{Compute reduced constraint coefficients}
\usage{
reduce_constr(data_constr_coef, tag_outcome_reduce, coef_reduce)
}
\arguments{
\item{data_constr_coef}{A dataframe where constraint coefficients are stored}

\item{tag_outcome_reduce}{Name of the tag_outcome to reduce}

\item{coef_reduce}{Numeric value to multiply the constraint coefficients of the selected tag_outcome}
}
\value{
A dataframe with tag_outcome with the updated coefficients.
}
\description{
Reduce constraint coefficient for one specific tag_outcome
}
\examples{
tag_outcomes <- c('energy', 'protein', 'ghge')
constr_coef_df <- set_constr_coef(tag_outcome = tag_outcomes, 
  coef_lwr = rep(0.9, length(tag_outcomes)), 
  coef_upr = rep(1.0, length(tag_outcomes)))
# reduce the coefficients of ghge to 90\% its original value
reduce_constr(data_constr_coef = constr_coef_df, 
  tag_outcome_reduce = 'ghge', 
  coef_reduce = 0.9)

}
