% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_results.R
\name{return_new_diet}
\alias{return_new_diet}
\title{Collect results from the new diet discovery}
\usage{
return_new_diet(result_obj, data_current_diet)
}
\arguments{
\item{result_obj}{Result object from the new diet discovry}

\item{data_current_diet}{Current diet that has food name and average intake}
}
\value{
A dataframe with food names, current diet and new diet
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
