% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_input.R
\name{select_diet}
\alias{select_diet}
\title{Select a set of foods}
\usage{
select_diet(data_diet, tag_food, minmax = T)
}
\arguments{
\item{data_diet}{Data that contain food (group) names and current diet information.}

\item{tag_food}{A vector of food names to choose from}

\item{minmax}{Whether to return min and max limit of feasible set for new diet}
}
\value{
A dataframe with food names and diet intake for selected foods
}
\description{
Food selector based on a bigger table, used to provide current diet information.
}
\examples{
foods <- c('Bread', 'Vegetables', 'Red meat')
select_diet(data_diet = all_diet, tag_food = foods)
}
