% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_input.R
\name{select_perunit}
\alias{select_perunit}
\title{Select contribution per unit for nutrition and environmental impact outcomes}
\usage{
select_perunit(data_perunit_contrib, tag_food, tag_outcome)
}
\arguments{
\item{data_perunit_contrib}{Data that contain food (group) names and current diet information.}

\item{tag_food}{A vector of food names to choose from}

\item{tag_outcome}{Whether to return min and max limit of feasible set for new diet}
}
\value{
A dataframe with food names and diet intake for selected foods
}
\description{
Select contribution per unit information based on a bigger table.
}
\examples{
foods <- c('Bread', 'Vegetables', 'Red meat')
tag_outcomes <- c('energy', 'protein', 'ghge')
select_perunit(data_perunit_contrib = contrib_per_unit, tag_food = foods, tag_outcome = tag_outcomes)
}
