% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_constraints.R
\name{set_constr_coef}
\alias{set_constr_coef}
\title{Set constraint coefficients}
\usage{
set_constr_coef(tag_outcome, coef_lwr, coef_upr)
}
\arguments{
\item{tag_outcome}{A vector of tag_outcome.}

\item{coef_lwr}{A vector of values to set the lower bound, used to multiply by total contribution.}

\item{coef_upr}{A vector of values to set the upper bound, used to multiply by total contribution.}
}
\value{
A dataframe with tag_outcome with the coefficients.
}
\description{
The constraint coefficients are used to compute the inequality constraints based on the total contribution.
The coefficients come in pairs: for each tag_outcome, there are lower and upper bounds for the constraint values.
}
\examples{
tag_outcomes <- c('energy', 'protein', 'ghge')
set_constr_coef(tag_outcome = tag_outcomes, 
  coef_lwr = rep(0.9, length(tag_outcomes)), 
  coef_upr = rep(1.0, length(tag_outcomes)))
}
