% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_results.R
\name{tab_contrib}
\alias{tab_contrib}
\title{Make \code{gt} table for the diet contribution summary}
\usage{
tab_contrib(tab_obj)
}
\arguments{
\item{tab_obj}{a table object from prep_contrib()}
}
\value{
a \code{gt} table with coloring
}
\description{
Make \code{gt} table for the diet contribution summary
}
\examples{
tab_contrib(tab_obj = dtt)
}
