% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_results.R
\name{table_validate}
\alias{table_validate}
\title{Make \code{gt} table for the diet contribution summary}
\usage{
table_validate(tab_obj)
}
\arguments{
\item{tab_obj}{a table object from prep_contrib()}
}
\value{
a \code{gt} table with coloring
}
\description{
Make \code{gt} table for the diet contribution summary
}
