% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_results.R
\name{validate_diet_contrib}
\alias{validate_diet_contrib}
\title{Compare new diet with current diet}
\usage{
validate_diet_contrib(data_new_diet, data_unit_contrib, data_constr)
}
\arguments{
\item{data_new_diet}{New diet intake}

\item{data_constr}{Constraint values to compare the new total contribution}

\item{data_current_diet}{Current diet that has food name and average intake}
}
\value{
A dataframe with tag_outcome, total contribution from the diet and whether it is within constraints
}
\description{
Compare new and current diet in terms of absolute and relative (percent) change
}
