% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_constraints.R
\name{values_by_tag_outcome}
\alias{values_by_tag_outcome}
\title{Organize input values for outcomes}
\usage{
values_by_tag_outcome(data_unit_contrib, data_constr)
}
\arguments{
\item{data_unit_contrib}{A dataframe where total contribution (raw or standardised) for each \code{tag_outcome} is stored}

\item{data_constr}{A dataframe of lower and upper constraint coefficients for each \code{tag_outcome}}
}
\value{
A list of outputs, organized by tag_outcome
}
\description{
The optimization algorithm requires three inputs per tag_outcome, such as 'energy':
\itemize{
\item contribution per unit (either standardized or raw) for each food
\item upper bound for total contribution
\item lower bound for total contribution
}
}
